ProjectCode/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.8
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.13.0
python-multipart==0.0.6
boto3==1.34.0